Refactors Upload specs into a separate file.
